+++ RELEASES.md
@@ -80,6 +80,7 @@ TBD
 - Repaired Bybit docs links in comment (#3125), thanks @sunlei
 - Repaired Bybit HTTP order place (#3127), thanks @sunlei
 - Repaired Bybit `AccountPosition` message parsing (#3147), thanks @sunlei
+- Repaired Bybit conditional order trigger semantics and type
 - Upgraded implied-vol crate (#3115), thanks @faysou
 - Upgraded Rust (MSRV) to 1.91.0
 - Upgraded Cython to v3.2.0